Cleaned up the scraper cog
