+++ cogs/scraper/wattpad_scraper.py
@@ -1,4 +1,10 @@
-from . import scraper
+"""Wattpad scraper cog.
+
+Many features to scrap from wattpad.
+
+"""
+from discord.ext import commands
+from cogs.utils import scraper
 
 #API_STORYINFO = 'https://www.wattpad.com/api/v3/stories/'
 #API_HOTSTORYLIST = 'https://www.wattpad.com/api/v3/stories?filter=hot'
@@ -6,6 +12,7 @@ API_NEWSTORYLIST = 'https://www.wattpad.com/api/v3/stories?filter=new'
 #API_STORYTEXT = 'https://www.wattpad.com/apiv2/storytext'
 #API_GETCATEGORIES = 'https://www.wattpad.com/apiv2/getcategories'
 #API_CHAPTERINFO = 'https://www.wattpad.com/apiv2/info'
+
 
 async def get_latest_story(session):
     story_json = (await scraper.get_json(session,
@@ -21,3 +28,28 @@ async def get_latest_story(session):
             .format(story_title,
                     story_description,
                     story_chapter_one_url))
+
+
+class WattpadScraper:
+    """Wattpad scraper cog."""
+
+    def __init__(self, bot):
+        """Initalization function."""
+        self.bot = bot
+
+    @commands.group(name='wattpad', pass_context=True)
+    async def _wattpad(self, ctx):
+        """Useful commands for getting stories from Wattpad."""
+        if ctx.invoked_subcommand is None:
+            await self.bot.pm_help(ctx)
+
+    @_wattpad.command()
+    async def latest(self):
+        """Get the latest story posted to Wattpad."""
+        result = await get_latest_story(self.bot.session)
+        await self.bot.say(result)
+
+
+def setup(bot):
+    """Called when cog is loaded via load_extension()."""
+    bot.add_cog(WattpadScraper(bot))
