+++ requirements.txt
@@ -1,3 +1,3 @@
-aiohttp==0.21.2
+aiohttp>=0.21.0,<0.22.0
 git+https://github.com/Rapptz/discord.py@async
 chatterbot
